Begin Vala migration with main/timespan_to_string
I hope in future to remove Glade and Granite entirely and code the UI
programmically in Vala.

This also cleans up the times_to_string function, which was two-thirds
code and one-third comment explaining the rationale for 17 bytes
allocated. This is removed in Vala (thank goodness).

activate(GApplication*, gpointer) is renamed so as not to cause a
conflict with the event name itself, which Vala errors about.