Clean up global variables and #defines
This should ease translation into Vala. It also seemed like a code smell
when I put them in, so I'm happy that they've been dealt with.

The only remaining globals are:
  * stillopen, which can probably be removed further into the Vala
    translation
  * set_time, which will probably become a static variable in Vala.

There is also widget_picker_pair introduced, which I hope to replace
with something along the lines of a tuple.